# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Untukmu API",
    "description": "Untukmu Fase 1–5: akun, pesan terenkripsi, sosial, doa, kurasi, pemulihan dan ekspor.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/constellations": {
      "get": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "List Galaxies",
        "operationId": "list_galaxies_api_v1_constellations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Create Galaxy",
        "operationId": "create_galaxy_api_v1_constellations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GalaxyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/constellations/{galaxy_id}": {
      "get": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Get Galaxy",
        "operationId": "get_galaxy_api_v1_constellations__galaxy_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Galaxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Patch Galaxy",
        "operationId": "patch_galaxy_api_v1_constellations__galaxy_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Galaxy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GalaxyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Delete Galaxy",
        "operationId": "delete_galaxy_api_v1_constellations__galaxy_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Galaxy Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages": {
      "post": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Create Message",
        "operationId": "create_message_api_v1_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/MessageInput"
                  },
                  {
                    "$ref": "#/components/schemas/PublicMessageInput"
                  },
                  {
                    "$ref": "#/components/schemas/UnlistedMessageInput"
                  }
                ],
                "title": "Body",
                "discriminator": {
                  "propertyName": "visibility",
                  "mapping": {
                    "private": "#/components/schemas/MessageInput",
                    "public_anon": "#/components/schemas/PublicMessageInput",
                    "unlisted": "#/components/schemas/UnlistedMessageInput"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/messages/{message_id}": {
      "get": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Get Message",
        "operationId": "get_message_api_v1_messages__message_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Edit Message",
        "operationId": "edit_message_api_v1_messages__message_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/MessageInput"
                  },
                  {
                    "$ref": "#/components/schemas/PublicMessageInput"
                  },
                  {
                    "$ref": "#/components/schemas/UnlistedMessageInput"
                  }
                ],
                "discriminator": {
                  "propertyName": "visibility",
                  "mapping": {
                    "private": "#/components/schemas/MessageInput",
                    "public_anon": "#/components/schemas/PublicMessageInput",
                    "unlisted": "#/components/schemas/UnlistedMessageInput"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Delete Message",
        "operationId": "delete_message_api_v1_messages__message_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/constellations/{galaxy_id}/messages": {
      "get": {
        "tags": [
          "Galaksi dan pesan privat"
        ],
        "summary": "Galaxy Messages",
        "operationId": "galaxy_messages_api_v1_constellations__galaxy_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Galaxy Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cipher": {
        "properties": {
          "v": {
            "type": "integer",
            "const": 2,
            "title": "V",
            "default": 2
          },
          "alg": {
            "type": "string",
            "const": "A256GCM",
            "title": "Alg",
            "default": "A256GCM"
          },
          "mode": {
            "type": "string",
            "const": "private",
            "title": "Mode",
            "default": "private"
          },
          "iv": {
            "type": "string",
            "title": "Iv"
          },
          "ct": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 24,
            "title": "Ct"
          },
          "aad": {
            "type": "string",
            "maxLength": 100,
            "title": "Aad"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "iv",
          "ct",
          "aad"
        ],
        "title": "Cipher"
      },
...
      "GalaxyInput": {
        "properties": {
          "target_kind": {
            "type": "string",
            "enum": [
              "orang_tua",
              "saudara",
              "sahabat",
              "pasangan",
              "seseorang",
              "diri_sendiri",
              "custom"
            ],
            "title": "Target Kind"
          },
          "target_label": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Target Label"
          },
          "custom_category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Category"
          },
          "visual_type": {
            "type": "string",
            "enum": [
              "photo",
              "light_symbol"
            ],
            "title": "Visual Type",
            "default": "light_symbol"
          },
          "visual_ref": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Ref"
          },
          "kind": {
            "type": "string",
            "enum": [
              "spiral",
              "ellipsoid",
              "irregular"
            ],
            "title": "Kind",
            "default": "spiral"
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Color",
            "default": "#ffd9a0"
          },
          "radius": {
            "type": "integer",
            "enum": [
              110,
              140,
              175
            ],
            "title": "Radius",
            "default": 140
          },
          "particle_count": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Particle Count",
            "default": 8500
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_kind",
          "target_label"
        ],
        "title": "GalaxyInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "constellation_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Constellation Ids"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "date_label": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Label"
          },
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Attachment Ids"
          },
          "visibility": {
            "type": "string",
            "const": "private",
            "title": "Visibility",
            "default": "private"
          },
          "payload": {
            "$ref": "#/components/schemas/Cipher"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "constellation_ids",
          "payload"
        ],
        "title": "MessageInput"
      },
...
      "PublicMessageInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "constellation_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Constellation Ids"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "date_label": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Label"
          },
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Attachment Ids"
          },
          "visibility": {
            "type": "string",
            "const": "public_anon",
            "title": "Visibility"
          },
          "public_body": {
            "type": "string",
            "maxLength": 40000,
            "minLength": 1,
            "title": "Public Body"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "constellation_ids",
          "visibility",
          "public_body"
        ],
        "title": "PublicMessageInput"
      },
...
      "UnlistedMessageInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "constellation_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Constellation Ids"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "date_label": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Label"
          },
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Attachment Ids"
          },
          "visibility": {
            "type": "string",
            "const": "unlisted",
            "title": "Visibility"
          },
          "public_body": {
            "type": "string",
            "maxLength": 40000,
            "minLength": 1,
            "title": "Public Body"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "constellation_ids",
          "visibility",
          "public_body"
        ],
        "title": "UnlistedMessageInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}